requirements.txt
flask
flask_sqlalchemy
flask_login
flask_bcrypt
pymysql
python-dotenv
cryptography
